feat: 스트레스 데이터 type 변경

See merge request s11-final/S11P31A408!62

+++ backend/api/src/main/java/com/scare/api/solution/walk/service/command/dto/SaveWalkingCourseDto.java
@@ -16,7 +16,7 @@ public class SaveWalkingCourseDto {
 	private double distance;
 	private LocalDateTime startedAt;
 	private LocalDateTime finishedAt;
-	private List<Integer> heartRates;
+	private List<Double> heartRates;
 	private SaveWalkingCourseStressDto stressData;
 	private List<SaveWalkingCourseLocationDto> locations;
